pyproject.toml
[project]
name = "django-boards-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "asgiref==3.11.0",
    "attrs==25.4.0",
    "Django==6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular==0.29.0",
    "inflection==0.5.1",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "pillow>=12.0.0",
    "PyYAML==6.0.3",
    "referencing==0.37.0",
    "rpds-py==0.30.0",
    "sqlparse==0.5.5",
    "tzdata==2025.3",
    "uritemplate==4.2.0",
]
